lambda/requirements.txt
psycopg2-binary==2.9.9
pika==1.3.2
PyJWT==2.8.0
bcrypt==4.1.2
boto3==1.34.0
